Fix endianess of BN arrays exported to OpenSSL
Apparently the order returned by the token is always big endian,
whiel OpenSSL fancies little endian in BN ...

Tested with:
openssl pkeyutl -verify -pubin -inkey "pkcs11:type=public" -rawin \
    -digest SHA256 -sigfile [sig file] -in [signed content]

Note that -pubin causes OpenSSL to tell the provider to just load the
public key. In this case we can easily export it, and that means OpenSSL
ends up taking the public key and importing it into the default provider
in the default configuration. This is how I know that the expot needed
reversing, as public signature verification was failing with:
[..]rsa_ossl_public_decrypt:data too large for modulus[..]
Which after a quick debugging turned out to just be luck that the least
significant bits of the modules where smaller when taken as most
significant bit and made me discover the whole endianess issue.

Signed-off-by: Simo Sorce <simo@redhat.com>