Add more tests and fix a very subtle bug
The last signature verification test in the patch was failing oddly.

Turn out that by calling the macro within an else block I was creating
a variable on the stack that, correctly, was recycled as soon as the
block was complete, releasing the stack where the value was stored.

The funny part was that this part of the stack was deep enough not to be
touched immediately, so other manual tests seemd to work just fine as
the memory passed into the callback was still untouched by cheer luck.

We need to keep that part of the stack allocated until the callback
returns. After that point we are assured the key has been imported and
the params memory can be released.

Do that by keeping the allocation of the "fix" buffers in the main
block so they cannot be release until the function completes, after the
callback is done.

Signed-off-by: Simo Sorce <simo@redhat.com>