Modify the key pointers in find_keys() only on success
The find_keys() function could initialize the *pub and *priv key
parameters while returning an error.  Fix the internal logic to only set
the *pub and *priv pointers on a success.  While it's normally ok to
expect the extra parameters to be in undefined state, at least in
p11prov_store_load(), it might happen that *pub and *priv would be used
later even if find_keys() returned error.

Signed-off-by: Ondřej Surý <ondrej@sury.org>