Allow store to enumerate objects
This allows to load multiple objects (currently only keys) and enumerate
them all.

Note: We need to store the login_session on the provider context because
otherwise we can't guarantee the token is "logged in" for the duration
of operations on private key later on.
We could create a session object and take a refeence on key that come
from logged in slots, but will leave that for later.

Issue that had to be wroked around: the password callback.
I used it to allow interactive pin usage, but the callabcks are only passed
to the _load() function.

Therefore we need code to defer loading or retry loading if a login was
required (basically always for private keys).

May be a bug to open with OpenSSL, UI pin callbacks should be provided
also at store open.

Signed-off-by: Simo Sorce <simo@redhat.com>