Add RSA PSS dispatchers and minor refactoring
We claim we can do RSA PSS in keymgmt.
We still do not support adding all pss restrictions as documented
in the openssl genpkey manpage for pss keys (digest, mfg, etc), as
there is no pkcs11 support for most of them.
We can restrict what digest type to use via allowed mechanisms, but
only for tokens that support setting the CKA_ALLOWED_MECHANISM
paramter on key objects, and even then we do not have a guarantee
that this translates to actual restrictions in the token.

Signed-off-by: Simo Sorce <simo@redhat.com>