CID 358270: Incorrect deallocator used
Coverity points out we are using OPENSSL_free() for a variabl (_name)
that shuld be freed by free() as it was allocated via strdup().
On the platforms we currently support OPENSSL_free() is just free() so
this is a minor issue, yet it could lead to bad memory corruption if an
application is able to change what allocators/deallocators are used and
this ends up in the provider context.
The fix here is to instead allocate it via OPENSSL_strdup().

Signed-off-by: Simo Sorce <simo@redhat.com>