Export public keys only
If export_object returns any data at all then openssl will often try to
use the default provider to perform operations as it assume the key is
exportable an re-importable to that provider.

So unless we are actually holding a public key return an error here so
that inside libcrypto the proper fallbacks can happen.

Signed-off-by: Simo Sorce <simo@redhat.com>