Require at least OpenSSL 3.0.5
There are fixes that landed in 3.0.5 that are required for proper
fallbacks in provider selection which we depend on in our test suite.
Although technically we can make the provider work by always passing a
"provider=pkcs11" property query, we do want to make sure most stuff
will work without needing it anyway.

Signed-off-by: Simo Sorce <simo@redhat.com>