Handle signature size request first
It is legal to call DigestSignFinal before any update operation has been
requested to know what the final signature size will be right away, just
base on the initialization which sets the algorithm type.
In this case there will be no session yet because that is requested on
demand when the first data is provided for the digest.

Signed-off-by: Simo Sorce <simo@redhat.com>