Add a way to trigger a covscan manually.
This allows to trigger a covscan on a pull request manually before
merging.

We only have a limited number of covscans we can run per day but we also
should run them before the PR merges to avoid followup fixes.
By using a lable to control when a covscan happens we can triger it just
before pressing the merge button and fix any found issues before the
merge.

Signed-off-by: Simo Sorce <simo@redhat.com>