Refator test suite
Use the setup function only once, once the setup is complete all the
needed variables for the tests are in the testvars, and make chck does
not rerun setup unless the temp dirs are removed.

Refactor tests to use a common wrapper that takes care of sourcing the
correct env vars based on which token they need to test.

The softhsm proxy is now spun of to a standalone wrapper that can be
used by individual tests as needed.

The test name now indicates what kind of token is used as well as
additional options needed (like the need for a proxy).

Signed-off-by: Simo Sorce <simo@redhat.com>