Fix "tests" when built outside source directory
The test now work when built in the source or some other directory
and configured like ../pkcs11-provider/configure

Define "testsblddir" and "TESTSBLDDIR" to point to the build/tests directory
i.e. "@abs_top_builddir@/tests" and correct several places where
original code made some assumptions that the source directory was the
same as the build directory.

Renamed the "testsdir" and "TESTSDIR" to "testssrcdir" and "TESTSSRCDIR"
to match the "testsblddir" and "TESTSBLDDIR" convention

`tests/test-wrapper` was patched to allow a file path name to include "-"
It would fail to run some of the tests because it would not find the
command an options.

Remove "export PINFILE=..."

Location of PINFILE is not needed outside the setup routines and even
if needed PINVALUE is exported and could be be used. Changes to be committed:

 Changes to be committed:
	modified:   tests/Makefile.am
	modified:   tests/openssl.cnf.in
	modified:   tests/setup-softhsm.sh
	modified:   tests/setup-softokn.sh
	modified:   tests/softhsm-proxy.sh
	modified:   tests/tbasic
	modified:   tests/teccsha2
	modified:   tests/test-wrapper
	modified:   tests/thkdf
	modified:   tests/toaepsha2
	modified:   tests/trsapss

 Date:     Mon Nov 21 14:52:24 2022 -0600

Signed-off-by: Doug Engert <deengert@gmail.com>