Refactor and merge P11PROV_OBJ and P11PROV_KEY
Keeping these objects separate was awkward as some places had to take
P11PROV_OBJ and then just immediately reference out a key object from
it only to use it for a couple of queries and then free the reference
taken.

By merging the two objects in a single structure we simplify some of
these operations.

This is in preparation for adding another object class (CKO_CERTIFIATE)
that requires different some data, but still has a lot in common with
key objects.

Signed-off-by: Simo Sorce <simo@redhat.com>