Enable the store to actually return certs
The DER formatted CKA_VALUE we already have is all that is needed to
return a cert to openssl, from there openssl just parses the asn.1 into
a X509 object and does not need interaction with the provider anymore.

This code does not support passing trust anchor data at this time.

Signed-off-by: Simo Sorce <simo@redhat.com>