CID 433457: Waiting while holding a lock
Coverity complains that we are busy waiting while holding the pool lock.
Initially I thought this was fine, but in a threaded program it is
actually the wrong thing to do. Not only because the resource will be
hogged by a single thread, but also because otherthreads may be
returning session if allowed to continue their operations.

This commit also restores code lost during some refactor or another
that was meant to close token sessions if a soft limit where surpassed
in order to minimize the number of cached sessions we keep open.

Fixes also CID 433458: Waiting while holding a lock

Signed-off-by: Simo Sorce <simo@redhat.com>