Add a simple code generator
Often we want to use macros to simplify writing boilerplate.
On one hand this is good as it avoid tedious, error-prone work.
OTOH it makes reading and debugging code harder when trying to simply
traverse complex code bases either with code indexers/taggers or even
just in debuggers that get often confused when the source code file does
not match the compiled code.

This simple generator allows to write macros to auto generate parts of
code and uses the compiler pre-processor to spit out actual readable
code that can be formatted by clang-format and committed to the tree.

Signed-off-by: Simo Sorce <simo@redhat.com>