Session locking fixes
In PR #156 an attempt to fix some minor locking issues resulted in a
much worse potential deadlock within the code that attempts to create a
login session.

The code roughly did the following:
    token_login {

        /* A section */
        lock -> pool
        lock -> session
        unlock -> pool

        /* issue  */

        /* B section */
        lock->pool
        unlock->pool
        return
    }

After section A, we hold the session lock, but do not hold the pool lock,
that means another thread could be calling token_login() before we enter
section B, and take the pool lock.
Then attempt to take the session lock and be stuck waiting.
At the same time the original thread will be stuck trying to get the pool
lock in section B, and not releaseing the session lock.
This is a classic deadlock situation and both threads will wait forever.

The basic solution is to use MUTEX_TRY_LOCK to get the session lock so that
in case of deadlock we'll get an error instead of being stuck forever.

After taking a step back however I reeliazed there were many shortcomings
in how the code was refactor to avoid waiting withing locked sections,
so the solution is a bit larger than a minimal fix, and better
streamlines some internal operations reducing the number of locking
sections to the bare minimum needed.

Signed-off-by: Simo Sorce <simo@redhat.com>