Handle authentication on token and auth failures
This patch improves the authantiction pathway.

It adds support for tokens that have an authentication device that does
not rely on the application to provide a PIN (like a physical pin pad, a
fingerprint scanner, etc...)

Add support for blocking bad pins from being reused multiple time
potentially causing the token to lock up.

Signed-off-by: Simo Sorce <simo@redhat.com>