Defer openening the store until first load
Unfortunately the current OpenSSL storege management interface will pass
in prompters only at load time, after "opening" the store.

Given we need to use the prompters at open data fetching time, we have
to defer the fetch operation otherwise only public objects will be
fetched.
This is a feature (we want to be able to fetch public objects w/o a
login), so the only way to do this properly for now, is to wait to do
anything until the user is actually prompted, and then try the
operation.

Signed-off-by: Simo Sorce <simo@redhat.com>