Fix asprintf usage
Use matched free function for slot->login_info

The current code calls OPENSSL_free on a memory region allocated with
asprintf, though OPENSSL_free may be replaced with a custom function
set with CRYPTO_set_mem_functions, which may be incompatible with the
system malloc.

This also fixes a memory leak and the check of return value as the
value of STRP argument is undefined if asprintf returns -1.

Signed-off-by: Daiki Ueno <dueno@redhat.com>