configure.ac: Fix libsoftokn3 detection on macOS
macOS uses .dylib as extension for shared libraries, but .so for
loadable modules. NSS seems to treat libsoftokn3 as shared library, so
it uses .dylib.

Autoconf has $SHARED_EXT for the platform-specific library suffix, so
use that. Add a comment above p11-kit-client.so indicating that this is
correct so we don't change it accidentally in the future.

See: #179
Signed-off-by: Clemens Lang <cllang@redhat.com>