Add fallback to digest + raw signature
Some tokens may implemnt the raw signature but not the combined digest
mechanisms. Work around thse lesser tokens by implementing the digest
calculation seprately, and only at the end calling a raw signature
function on the calculated digest.

When we fall back, we try to use pkcs11's own digest function in
preference and if they are available we also try to make sure they are
from the same slot that will later perform the signature operation.

Signed-off-by: Simo Sorce <simo@redhat.com>