Retry store_fetch with login if nothing found
This is to account for the case where HSMs may require login even for
public objects. They will return no objects at all unless there is a
login session.

If a user is sure their token always allow public objects to be
retrieved w/o PIN there is a tunable that can be set to avoid the
fallback annoyance when the pkcs11 uri is imply not matching a public
key correctly.

Signed-off-by: Simo Sorce <simo@redhat.com>