Clarify specific bit of obscure code
This code fooled myself as well in #212, so I think it needs some more
comment that makes it clearer what is going on here. The code is
confusing because it depends on interactions with the caller (which can
call repeatedly to move "forward" from slot to slot) and because of the
double anded negative condition, which is a bit hard to grok at first sight.

Signed-off-by: Simo Sorce <simo@redhat.com>