Rework session locking
In oreder to properly support resetting status after forking we need to
stop holding mutexes for unbounded time. Instead use mutexes only to
guard a variable that indicates is a session is in use or not.

This also simplifies the code that deal with login session quite as it
does not reuire anymore to cycle on try_lock because the code is never
expected to hold session locks indefinitely and we can fully serialize.

Signed-off-by: Simo Sorce <simo@redhat.com>