Require login when looking for asociated objects
Generally either we are looking for the public key or cert associated to
a private key (therefore we already must have had a login sessions).
Or we are trying to refresh keys to use after a fork event caused a
reset of objects.
In both cases private keys are invovled, so requiring a login is fine.
If we don't this action can fail to refresh objects if there is no other
operation that by chance required a login before the refresh was hit.

Signed-off-by: Simo Sorce <simo@redhat.com>