CID 451133: Fix errors using config option names
This error is not going to cause issues in practice as the structure
currently only contains a string anyway, so the address of the structure
does in fact correspond to the address of the string we want to print.
But in future may return garbage if the structure is changed.

Signed-off-by: Simo Sorce <simo@redhat.com>