Smarten up handling of login session
Although we want to ensure a login session is available for certain
operations, we also do not want to unnecessarily fail them when mutliple
operations are running concurrently.
As login status is token global, all we need to ensure is that a login
session exist. This can be reasonably easily determined by session
status as well.
So where we do not actually need to use the login_session directly we
can simply check the status and return positively.

The only place that really wants to use a login_session directly is the
caching layer which is tolerant to failure already.

Signed-off-by: Simo Sorce <simo@redhat.com>