Fix enable cross compilation by removing the AC_CHECK_FILE macro
From the man-page on the `AC_CHECK_FILE` macro, we can read that:

```
You might also need to check for the existence of files. Before using these macros, ask yourself whether a runtime test might not be a better solution. Be aware that, like most Autoconf macros, they test a feature of the host machine, and therefore, they die when cross-compiling.
```

Therefore, move to using a simple shell if/else as the replacement.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>