Fix softokn detection on macOS
Loadable modules on macOS should use .so as extension, but both NSS
softokn and OpenSSL providers do not follow this rule. Modify shlext
accordingly so that NSS softokn can be found for the tests.

Note that p11-kit correctly names its loadable modules with .so, so this
variable should be used carefully as it may be incorrect depending on
your use case.

Signed-off-by: Clemens Lang <cllang@redhat.com>