Make provider preferential, not mandatory
Apparently Ubuntu has some configuration that tries hard to load a
rdrand engine. When the propquery is set to a hard provider=pkcs11 this
fails as the tpe of DRBG being sourced becomes incompatible with the
mandate property.
Soften the test to only prefer the provider so that the operations we
care for will come from the pkcs11 provider (we check errors anyway) and
we do not get the noise of unrelated failures.

Signed-off-by: Simo Sorce <simo@redhat.com>