Allow also Raw PSS mechanisms in the default set
When an application pre-hashes the content to be signed it can use the
raw CKM_RSA_PKCS_PSS mechanism to apply a signature. This may be done
with simple hardware tokens that do not support digest operations on
board and need to rely on the software to deal with that part.
We should not preclude such use for key we generate.

Signed-off-by: Simo Sorce <simo@redhat.com>