Add option to block specific operations
This allows to configure pkcs11-provider to never expose a specific set
of operations.

This is so either bugs on the OpenSSL side (like assuming all digest
context can always be duplicated) or bugs on the token side (like making
a non-functional mechanism erroneously available) can be handled by
letting openssl take care of those operations with the default provider
when properties otherwise would force openssl to use the provider's
operations.

Signed-off-by: Simo Sorce <simo@redhat.com>