tests: Fix ASAN build on macOS
On macOS, the variable for library preloading is DYLD_INSERT_LIBRARIES,
not LD_PRELOAD. Adjust the CHECKER env variable to use that on macOS
machines.

Change fake_dlclose to a library rather than a shared module. This
doesn't make a difference on Linux, but on macOS, dyld(1) refuses to
insert a shared module into a processes address space using
DYLD_INSERT_LIBRARIES, but allows a library in the same place.

Signed-off-by: Clemens Lang <cllang@redhat.com>