Add support to store public key as session objects
This allows operations that actually need to happen on a token with
a stored key (unlike ECDH which just needs a buffer passed int), to
be able to actually import the object as an ephemeral session object
to perform the required operations.

We generally do not want to store public key to avoid littering the
login session (which is long lived) with too many throw-away objects
as we have no way to determine when it is ok to remove session objects
for long lived programs.

Signed-off-by: Simo Sorce <simo@redhat.com>