Fix memory leaks when tokens are missing
In case we have slots advertized but the driver fails to return
information or the token is not present we were leaking memory as the
slot is not added to the array and the number of slots is not
incremented.

Ensure the slot struct is freed in this case.
And ensure the slot is assigned and count incremented at the same time
to avoid leaks from other error conditions.

Signed-off-by: Simo Sorce <simo@redhat.com>