provider: Add new configuration option assume_fips
When OpenSSL runs in FIPS Mode, it will not use any providers
that do not provide a property fips=yes, rendering the pkcs11
provider unusable in FIPS Mode. This is a regression and for
many users that need to have smart cards working in FIPS Mode.

Unfortunately, proper signalization from pkcs11 modules regarding
the tokens FIPS certification status is not standardized yet,
this will be left up to the user to decide if the pkcs11 modules
talk to FIPS certified token or not.

This involves adjusting the algorithm lists to contain dynamic
properties based on this configuration option, where we previously
had hardcoded just provider=pkcs11.

Fixes: #469, #164

Signed-off-by: Jakub Jelen <jjelen@redhat.com>