tests: Factor out cert selfsigning to helper
At this moment, this was already copied into two places.

This also generates second restricted RSA-PSS key during setup to avoid the
need to generate key as part of rsapssam test, which makes it reentrant and
should make the uri test a bit faster as we will have less keys.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>