Keep track of associated objects
In a few case we need to keep track of associated objects in order to be
compatible with the way OpenSSL deals with key pairs. The EVP_PKEY
object can reference both a private and a public key at the same time.

Storing an "associated object" (generally a public key associated to the
private key) allows the code to perform operations wth a public key
where the EVP_PKEY primary object is actually a private key.

The key pair generation code now always stores the public key as an
associated object so that generating a keypair and then immediately
testing it with a pair of sign/verify operations can be accomplished.

Signed-off-by: Simo Sorce <simo@redhat.com>