Ensure we correctly deal with import selection
In the ECDH case import involves these steps:

- create new empty key
- set EC domain parameters, by copying them from the server key
- set EC Public key from the peer

The previous code was messing the second step by completely copying the
server public key if one was found on the system.

This is because both on export and import we did not look at the selection
and always exported/imported both parameters and the public key.

As we now retain the public key object we generate with keygen, instead
of discarding it, the export function would export also the public point
and the import would find the server public key cached and import it
fully in the empty shell reservd for the public key.

Later setting the peer public key would then fail as the object was
considered a complete key already.

Signed-off-by: Simo Sorce <simo@redhat.com>