Drop SoftHSM from the matrix for now
We build on Mac OS X just to ensure the code builds, however we do not
absolutely need to run multiple tests, as functionality with the various
tokens is already tested in the rest of the CI

Given lately MAc OS X 14 + SoftHSM always times out, disable it from the
matrix for now.

Note that the rest of the conditionals are intentionally preserved, so
that we can simply re-add sofhtsm to the matrix should we want to.

Signed-off-by: Simo Sorce <simo@redhat.com>