Addjust tests to deal with ASAN
Changing openssl.cnf early causes pkcs11-tool to blow up as it loads
a shared modules linked to libasan and the address sanitizer doesn't
like that.

Note that we can't simply preload libasan here because doing that causes
a not well understood segfault in pkcs11-tool at this time, so we just
run pkcs11-tool via a helper that prevents access to the tests
openssl.cnf

Additionally run openssl command via the helper that actually correctly
preloads libasan like doen in the ret of the tests.

Signed-off-by: Simo Sorce <simo@redhat.com>