objects: Set quirk only when the attribute is not on private key
The usage of the ALLOWED_MECHANISMS on public keys is quite limited/questionable
so some tools (notably pkcs11-tool before 0.27.0 [1]) was setting this attribute
only on the private keys.

When in some workflow, the first loaded key is a public key without allowed
mechanisms, the following queries for private keys skip reading of this
attribute, which can lead to incompatible keys and certificates.

This will still be an issue for longer-running programs that will load first key
without this attribute (for example ECC) and then it will expect the RSA-PSS
key.

The return value CKR_ATTRIBUTE_TYPE_INVALID for GetAttribute does not denote the
token does not know this attribute, but that this key does not have this
attribute.

The other option might be changing kryoptic to not return this error and do like
SoftHSM, returning empty list (which has different drawbacks).

[1] https://github.com/OpenSC/OpenSC/pull/3344

Signed-off-by: Jakub Jelen <jjelen@redhat.com>