object: Fix ML-DSA keys comparison
Previously, each of the key pair had different size defined, which broke the
comparsion. It also missed the implementation of the public&private key
comparison altogether.

OpenSSL documentation is sparse [1], but I got a clarification that the bit size
here should really by the bit size of the public key.

[1] https://github.com/openssl/openssl/issues/28337

Signed-off-by: Jakub Jelen <jjelen@redhat.com>