Remove open_sessions counter from session pool
The `open_sessions` counter in the session pool was prone to accounting errors
and race conditions. It was difficult to keep the count synchronized,
especially if sessions were closed by external events.

This change removes the counter entirely. Instead, when a session is returned
to the pool, the code now dynamically counts the number of currently open
sessions.

This new approach is more robust and simplifies the session management logic
by eliminating the need for locking in many of the fucntions.

Note that the counting is still not necessarily accurate, the handle
stored in any session can change while we are counting causing
potentially some undercointing or over counting. And multiple session
releases can race and end up over counting te number of open sessions.

However closing more session than the max allowed is not a problem. With ref
counting we prevent unecesary closure of sessions that are holding data so no
harm is done.

Signed-off-by: Simo Sorce <simo@redhat.com>