Avoid using the login session when not necessary
Taking a login session should be sone only when there are no other
options. Now that objects can reference sessions and guarantee
they are not closed while the object is still in use avoid taking
the login session itself as the reference will cause increase use
of this session via p11prov_try_session_ref(), which in turn may
cause access to this session to fail due to concurrent use.

Signed-off-by: Simo Sorce <simo@redhat.com>