Add workaround for tokens that fail cleanup
Some PKCS#11 modules, notably NSS softokn, have a broken interface and may
fail to terminate an ongoing operation when the respective Init function
(e.g., C_EncryptInit) is called with a NULL mechanism. This can result in a
CKR_MECHANISM_PARAM_INVALID error, leaving the session in an unusable state.

To improve robustness, this change introduces a fallback mechanism. If the
initial attempt to terminate an operation fails, the provider will now attempt
to use the PKCS#11 v3.0 C_SessionCancel function.

If both termination methods fail, the session is considered unrecoverable. It
is marked as "broken" and taken out of the session pool to prevent it from
being used for subsequent operations. The session will be closed and recycled
once it is no longer in active use.

Signed-off-by: Simo Sorce <simo@redhat.com>