Sig: find an associated public key if private
OpenSSL often assumes EVP_PKEYs are keypairs and specifically that
private keys always have their associated public key available or
derivable.

One case is CSR signign where the code uses the private key to verify
the signature expecting it to have the public pair with it.

Verification initialization already used p11prov_obj_get_associated() as
a fallback case, but this function does not search the public key, only
works if a public key was already associated. This covers situations
where the key was generated as part of CSR creation, because keygen
always associates the corresponding public key. But it does not cover
the case where a pre-existing key is used. By switching to find we cover
all cases.

Signed-off-by: Simo Sorce <simo@redhat.com>